task3 elite/requirement.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
Pillow>=9.4.0
